+++ packages/compute-app/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @selva/compute-app
+# selva-compute-app
 
 ## 0.3.0
 
 ### Patch Changes
 
 - Updated dependencies [cd6ad4b]
-  - @selva/compute@1.2.0
+  - selva-compute@1.2.0
   - @selva/shared@1.2.0
 
 ## 0.2.0
@@ -20,5 +20,5 @@
   - Improve error handling and layout structure
 - Updated dependencies
 - Updated dependencies
-  - @selva/compute@1.1.0
+  - selva-compute@1.1.0
   - @selva/shared@2.0.0